% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{acceptable_set}
\alias{acceptable_set}
\title{Doses at least as close to the target as the MTD}
\usage{
acceptable_set(pi, phi = 0.25)
}
\arguments{
\item{pi}{True DLT probabilities.}

\item{phi}{Target toxicity rate.}
}
\value{
Integer vector of dose indices.
}
\description{
Doses at least as close to the target as the MTD
}
