% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_decomposition}
\alias{fig_decomposition}
\title{Decomposition of accuracy differences relative to BOIN}
\usage{
fig_decomposition(df)
}
\arguments{
\item{df}{Data frame from \code{\link{summarise_results}}.}
}
\value{
A ggplot object.
}
\description{
Decomposition of accuracy differences relative to BOIN
}
