% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_elim}
\alias{fig_elim}
\title{Elimination rule comparison}
\usage{
fig_elim(elim_df)
}
\arguments{
\item{elim_df}{Data frame from \code{\link{elim_rule_study}}.}
}
\value{
A ggplot object.
}
\description{
Elimination rule comparison
}
