% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_loss}
\alias{fig_loss}
\title{Overdose-aversion frontier}
\usage{
fig_loss(loss_res)
}
\arguments{
\item{loss_res}{List from \code{\link{loss_sensitivity}}.}
}
\value{
A ggplot object.
}
\description{
Overdose-aversion frontier
}
