% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_nsens}
\alias{fig_nsens}
\title{Accuracy against sample size}
\usage{
fig_nsens(n_res, scs = c(6, 9, 1, 5))
}
\arguments{
\item{n_res}{List from \code{\link{n_sensitivity}}.}

\item{scs}{Scenario indices to show.}
}
\value{
A ggplot object.
}
\description{
Accuracy against sample size
}
