% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_pcs}
\alias{fig_pcs}
\title{Accuracy by design and scenario}
\usage{
fig_pcs(df)
}
\arguments{
\item{df}{Data frame from \code{\link{summarise_results}}.}
}
\value{
A ggplot object.
}
\description{
Accuracy by design and scenario
}
