% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_steep}
\alias{fig_steep}
\title{Steepness gradient}
\usage{
fig_steep(steep_df)
}
\arguments{
\item{steep_df}{Data frame from \code{\link{steep_curve_study}}.}
}
\value{
A ggplot object.
}
\description{
Steepness gradient
}
