% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_straddle}
\alias{fig_straddle}
\title{Offset-MTD behaviour across sample sizes}
\usage{
fig_straddle(str_df)
}
\arguments{
\item{str_df}{Data frame from \code{\link{straddle_N_study}}.}
}
\value{
A ggplot object.
}
\description{
Offset-MTD behaviour across sample sizes
}
