% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{is_tie}
\alias{is_tie}
\title{Does a scenario contain an exact tie?}
\usage{
is_tie(pi, phi = 0.25)
}
\arguments{
\item{pi}{True DLT probabilities.}

\item{phi}{Target toxicity rate.}
}
\value{
Logical.
}
\description{
Does a scenario contain an exact tie?
}
