% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_crm}
\alias{table_crm}
\title{CRM skeleton sensitivity table}
\usage{
table_crm(crm_res)
}
\arguments{
\item{crm_res}{List from \code{\link{crm_sensitivity}}.}
}
\value{
A data frame.
}
\description{
CRM skeleton sensitivity table
}
