% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_safety}
\alias{table_safety}
\title{Safety operating characteristics}
\usage{
table_safety(df, designs = primary_designs)
}
\arguments{
\item{df}{Data frame from \code{\link{summarise_results}}.}

\item{designs}{Design names.}
}
\value{
A list of three wide data frames: overdose selection, mean DLTs, and
  the percentage of patients treated above the MTD.
}
\description{
Safety operating characteristics
}
